Fix a race condition delay when waiting on flex counter threads to exit
In some cases, it's possible that when the request to stop a flex
counter thread is received, the thread is in the middle of a sleep. This
adds unnecessary delay.

Fix this by making sure we have the sleep lock when we send out
notifications via both condition variables, and modify the wait and
wait_for calls to check to see if the thread needs to still run.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>